rename main.py to run.py, app dir to pyfastapi, then refactor the imports to reflect the name change
