+++ alembic/env.py
@@ -1,15 +1,10 @@
-from dotenv import load_dotenv
-load_dotenv()
-
 from logging.config import fileConfig
-
 from sqlalchemy import engine_from_config
 from sqlalchemy import pool
-
 from alembic import context
 
-from app.config import config as app_config
-print(app_config)
+from pyfastapi.config import config as app_config
+
 # this is the Alembic Config object, which provides
 # access to the values within the .ini file in use.
 config = context.config
