+++ tests/api_tests/test_continents.py
@@ -1,6 +1,6 @@
 from fastapi.testclient import TestClient
 
-from app.main import app
+from pyfastapi.main import app
 
 client = TestClient(app)
