+++ tests/test_main.py
@@ -1,6 +1,6 @@
 from fastapi.testclient import TestClient
 
-from app.main import app
+from pyfastapi.main import app
 
 client = TestClient(app)
